separate alignment module and edit protein start
